+++ src/helpers/transferBalance.ts
@@ -1,7 +1,16 @@
 import { ApiPromise } from '@polkadot/api'
-import { AccountId } from '@polkadot/types/interfaces'
-import { Callback, IKeyringPair, ISubmittableResult } from '@polkadot/types/types'
-import { BN } from '@polkadot/util'
+import { SubmittableExtrinsicFunction } from '@polkadot/api/types'
+import { AccountId, EventRecord } from '@polkadot/types/interfaces'
+import { AnyTuple, Callback, IKeyringPair, ISubmittableResult } from '@polkadot/types/types'
+import { BN, bnToBn } from '@polkadot/util'
+import { checkIfBalanceSufficient } from './checkIfBalanceSufficient'
+import { ExstrinsicThrowErrorMessage, getExtrinsicErrorMessage } from './getExtrinsicErrorMessage'
+
+export type TransferBalanceResult = {
+  result?: ISubmittableResult
+  errorMessage?: ExstrinsicThrowErrorMessage | 'ExtrinsicFailed'
+  errorEvent?: EventRecord
+}
 
 /**
  * Transfers a given amount of tokens from one account to another.
@@ -10,14 +19,24 @@ export const transferBalance = async (
   api: ApiPromise,
   fromAccount: IKeyringPair | string,
   toAddress: string | AccountId,
-  amount: BN,
+  amount: bigint | BN | string | number,
+  allowDeath?: boolean,
   statusCb?: Callback<ISubmittableResult>,
-) => {
+): Promise<TransferBalanceResult> => {
+  const hasSufficientBalance = await checkIfBalanceSufficient(api, fromAccount, amount)
+  if (!hasSufficientBalance) {
+    return Promise.reject({ errorMessage: 'TokenBelowMinimum' } satisfies TransferBalanceResult)
+  }
+
   return new Promise(async (resolve, reject) => {
     try {
-      const unsub = await api.tx.balances
-        .transfer(toAddress, amount)
-        .signAndSend(fromAccount, (result) => {
+      const transferFn = (api.tx.balances[
+        allowDeath ? 'transferAllowDeath' : 'transferKeepAlive'
+      ] || api.tx.balances['transfer']) as SubmittableExtrinsicFunction<'promise', AnyTuple>
+
+      const unsub = await transferFn(toAddress, bnToBn(amount)).signAndSend(
+        fromAccount,
+        (result: ISubmittableResult) => {
           statusCb?.(result)
           const isInBlock = result?.status?.isInBlock
           if (!isInBlock) return
@@ -26,20 +45,21 @@ export const transferBalance = async (
           )
           if (isInBlock && errorEvent) {
             // Reject if `ExtrinsicFailed` event was found
-            reject({
-              errorMessage: 'ExtrinsicFailed',
-              errorEvent,
-            })
+            reject({ errorMessage: 'ExtrinsicFailed', errorEvent } satisfies TransferBalanceResult)
             unsub?.()
           } else if (isInBlock) {
             // Otherwise resolve succesfully if transaction is in block
             resolve({ result })
             unsub?.()
           }
-        })
-    } catch (e) {
-      // Reject if user cancelled with `UserCancelled`
-      reject({ errorMessage: 'UserCancelled' })
+        },
+      )
+    } catch (e: any) {
+      console.error('Error while transferring balance:', e)
+      reject({
+        errorMessage: getExtrinsicErrorMessage(e),
+        errorEvent: e,
+      } satisfies TransferBalanceResult)
     }
   })
 }
@@ -53,12 +73,17 @@ export const transferFullBalance = async (
   toAddress: string | AccountId,
   keepAlive?: boolean,
   statusCb?: Callback<ISubmittableResult>,
-) => {
+): Promise<TransferBalanceResult> => {
+  const hasSufficientBalance = await checkIfBalanceSufficient(api, fromAccount)
+  if (!hasSufficientBalance) {
+    return Promise.reject({ errorMessage: 'TokenBelowMinimum' } satisfies TransferBalanceResult)
+  }
+
   return new Promise(async (resolve, reject) => {
     try {
       const unsub = await api.tx.balances
         .transferAll(toAddress, !!keepAlive)
-        .signAndSend(fromAccount, (result) => {
+        .signAndSend(fromAccount, (result: ISubmittableResult) => {
           statusCb?.(result)
           const isInBlock = result?.status?.isInBlock
           if (!isInBlock) return
@@ -67,10 +92,7 @@ export const transferFullBalance = async (
           )
           if (isInBlock && errorEvent) {
             // Reject if `ExtrinsicFailed` event was found
-            reject({
-              errorMessage: 'ExtrinsicFailed',
-              errorEvent,
-            })
+            reject({ errorMessage: 'ExtrinsicFailed', errorEvent } satisfies TransferBalanceResult)
             unsub?.()
           } else if (isInBlock) {
             // Otherwise resolve succesfully if transaction is in block
@@ -78,10 +100,12 @@ export const transferFullBalance = async (
             unsub?.()
           }
         })
-    } catch (e) {
-      console.error(e)
-      // Reject if user cancelled with `UserCancelled`
-      reject({ errorMessage: 'UserCancelled' })
+    } catch (e: any) {
+      console.error('Error while transferring full balance:', e)
+      reject({
+        errorMessage: getExtrinsicErrorMessage(e),
+        errorEvent: e,
+      } satisfies TransferBalanceResult)
     }
   })
 }